+++ src/components/Footer.js
@@ -11,7 +11,7 @@ export default React.createClass({
           <div className="ui grid">
             <div className="seven wide column">
               <div className="footer-box" style={{fontSize: '18px', paddingTop: '3rem'}}>
-                wowAfrica 阿非卡為台灣最完整的非洲資訊平台 最新資訊及活動通知我們的願景是搭起台灣與非洲的橋樑，藉由資訊、新聞、整合各方資源及實體活動等推廣方式，豐富我們對於非洲的認識及想像，達成進一步的交流與發展。
+                wowAfrica 阿非卡為台灣最完整的非洲資訊平台，我們的願景是搭起台灣與非洲的橋樑，藉由資訊、新聞、整合各方資源及實體活動等推廣方式，豐富我們對於非洲的認識及想像，達成進一步的交流與發展。
               </div>
             </div>
             <div className="four wide column" style={{color: 'white', backgroundColor: '#BF3333'}}>
@@ -23,16 +23,21 @@ export default React.createClass({
                   </a>
                 </div>
                 <div>
+                  <a href="/contact_us" onClick={this._onClick}>
+                     聯絡我們
+                  </a>
                 </div>
-                <div style={{paddingTop: '10px'}}>
+                <div>
+                </div>
+                <div className="footer-icon-groups">
                   <a href="mailto:lighteningdark2014@gmail.com" target="_blank">
-                    <img src="/images/rb_mail.png" style={{width: '50px', border: 'thin solid white', margin: '3px'}}/>
+                    <div className="footer-icon footer-icon-mail"/>
                   </a>
                   <a href="http://wowafrica.tumblr.com/" target="_blank">
-                    <img src="/images/rb_tumblr.png" style={{width: '50px', border: 'thin solid white', margin: '3px'}}/>
+                    <div className="footer-icon footer-icon-tumblr"/>
                   </a>
                   <a href="https://www.facebook.com/wowafrica.tw/" target="_blank">
-                    <img src="/images/rb_facebook.png" style={{width: '50px', border: 'thin solid white', margin: '3px'}}/>
+                    <div className="footer-icon footer-icon-facebook"/>
                   </a>
                 </div>
               </div>
@@ -40,12 +45,17 @@ export default React.createClass({
             <div className="five wide column">
               <div className="footer-box">
                 最新資訊及活動通知
-                <div className="ui mini input">
-                  <input id="emailInput" type="text" placeholder="Email"/>
+                <div id="emailForm" className="ui form" style={{paddingTop: '8px'}}>
+                  <div className="field">
+                    <input id="emailInput" type="email" placeholder="Email"/>
+                  </div>
+                  <div className="ui error message" style={{fontSize: '14px'}}>
+                    oops! 現在有些問題，請稍候再試
+                  </div>
+                  <div id="footer-emailBtn" className="ui submit button" onClick={this._onSubBtnClick}>
+                    訂閱
+                  </div>
                 </div>
-                <button id="emailBtn" className="ui button" onClick={this._onSubBtnClick}style={{backgroundColor: '#BF3333', color: 'white'}}>
-                  訂閱
-                </button>
               </div>
             </div>
           </div>
@@ -62,11 +72,21 @@ export default React.createClass({
   },
 
   _onSubBtnClick(e) {
+    $('#emailForm').removeClass('error');
+    $('#footer-emailBtn').removeClass('button').addClass('loading button');
     let request = $.ajax({
       url: 'https://script.google.com/macros/s/AKfycbzuOLvCRul3ZAzNl3kp9nPihUQ_iowpjB-Uf1nwuvhY9Q5lODI/exec',
       type: 'post',
-      data: {email: $('#emailInput')[0].value}
+      data: {'email': $('#emailInput')[0].value}
+    })
+    .done(function(data) {
+      $('#footer-emailBtn').text('完成!');
+    })
+    .fail(function(data) {
+      $('#emailForm').addClass('error');
+    })
+    .always(function() {
+      $('#footer-emailBtn').removeClass('loading button').addClass('button');
     });
-    $('#emailBtn')[0].innerText = '完成!';
   }
 });